integrates micro-services, rename shared model to shared contracts
