+++ Services/Notification/Troupon.Api.Notification/Startup.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
-using Troupon.Shared.Model;
+using Troupon.Shared.Contracts;
 
 namespace Troupon.Api.Notification
 {
