+++ Services/Ordering/Troupon.Core.Application.Ordering/Troupon.Core.Application.Ordering.csproj
@@ -17,7 +17,7 @@
     <ProjectReference Include="..\..\..\Infrastructure\Infra.MediatR\Infra.MediatR.Caching\Infra.MediatR.Caching.csproj" />
     <ProjectReference Include="..\..\..\Infrastructure\Infra.MediatR\Infra.MediatR.Events\Infra.MediatR.Events.csproj" />
     <ProjectReference Include="..\..\..\Infrastructure\Infra.Persistence\Infra.Persistence\Infra.Persistence.csproj" />
-    <ProjectReference Include="..\..\Shared\Model\Troupon.Shared.Model.csproj" />
+    <ProjectReference Include="..\..\Shared\Model\Troupon.Shared.Contracts.csproj" />
     <ProjectReference Include="..\Troupon.Core.Domain.Ordering\Troupon.Core.Domain.Ordering.csproj" />
     <ProjectReference Include="..\Troupon.Infra.Persistance.Ordering\Troupon.Infrastructure.Ordering.csproj" />
   </ItemGroup>
