+++ Services/Payment/Troupon.Api.Payment/Controllers/PaymentController.cs
@@ -1,9 +1,5 @@
 using System.Collections.Generic;
-using System.Threading.Tasks;
-using MassTransit;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Troupon.Shared.Model;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
