--- Services/Payment/Troupon.Api.Payment/Events/ApprovedOrderEvent.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Troupon.Api.Payment.Events
-{
-  public class ApprovedOrderEvent
-  {
-     public Guid OrderId { get; set; }
-  }
-}
