+++ Services/Payment/Troupon.Api.Payment/Troupon.Api.Payment.csproj
@@ -18,8 +18,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\Infrastructure\Infra.Common.Models\Infra.Common.Models.csproj" />
-    <ProjectReference Include="..\..\Shared\Model\Troupon.Shared.Model.csproj" />
+    <ProjectReference Include="..\..\Shared\Model\Troupon.Shared.Contracts.csproj" />
     <ProjectReference Include="..\Troupon.Core.Application.Payment\Troupon.Core.Application.Payment\Troupon.Core.Application.Payment.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Events" />
   </ItemGroup>
 
 </Project>
