+++ Services/Payment/Troupon.Core.Application.Payment/Troupon.Core.Application.Payment/Troupon.Core.Application.Payment.csproj
@@ -14,7 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\Infrastructure\Infra.DomainDrivenDesign.Base\Infra.DomainDrivenDesign.Base.csproj" />
-    <ProjectReference Include="..\..\..\Shared\Model\Troupon.Shared.Model.csproj" />
+    <ProjectReference Include="..\..\..\Shared\Model\Troupon.Shared.Contracts.csproj" />
   </ItemGroup>
 
   <ItemGroup>
