+++ Services/Shared/Model/Events/GuestOrderPlacedEvent.cs
@@ -3,8 +3,9 @@ using Infra.Common.Models;
 using Infra.DomainDrivenDesign.Base;
 using MediatR;
 
-namespace Troupon.Shared.Model.Contracts
+namespace Troupon.Shared.Contracts.Events
 {
+  //Order placed by a user guest event published by OrderingAPI and consumed by PaymentAPI
   public class GuestOrderPlacedEvent: IDomainEvent, IAuditable, INotification
   {
     public Guid OrderId { get; set; }
