+++ Services/Shared/Model/Events/OrderExecutionStartedEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace Troupon.Shared.Contracts.Events
+{
+  public class OrderExecutionStartedEvent
+  {
+    
+  }
+}
