+++ Services/Shared/Model/Troupon.Shared.Contracts.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Troupon.Shared.Model</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
