+++ Services/Shipment/Troupon.Api.Shipment/Troupon.Api.Shipment.csproj
@@ -5,9 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="MassTransit" Version="7.2.3" />
     <PackageReference Include="MassTransit.AspNetCore" Version="7.2.3" />
     <PackageReference Include="MassTransit.Extensions.DependencyInjection" Version="7.2.3" />
-    <PackageReference Include="MassTransit.RabbitMQ" Version="7.3.0-develop.293" />
+    <PackageReference Include="MassTransit.RabbitMQ" Version="7.2.3" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
