--- Services/Shipment/Troupon.Core.Application.Shipment/Events/OrderExecutionStartedEvent.cs
@@ -1,7 +1,0 @@
-﻿namespace Troupon.Core.Application.Shipment.Events
-{
-  public class OrderExecutionStartedEvent
-  {
-
-  }
-}
