+++ Services/Shipment/Troupon.Core.Application.Shipment/Troupon.Core.Application.Shipment.csproj
@@ -15,7 +15,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\Shared\Model\Troupon.Shared.Model.csproj" />
+      <ProjectReference Include="..\..\Shared\Model\Troupon.Shared.Contracts.csproj" />
     </ItemGroup>
 
 </Project>
